src/utils/cache.py
"""
Redis cache wrapper (optional, for future use).
Currently using in-memory storage, can be extended with Redis.
"""
from typing import Optional, Dict, Any
import structlog

logger = structlog.get_logger()


class Cache:
    """
    Simple cache interface.
    
    Currently in-memory only, can be extended with Redis.
    """
    
    def __init__(self, redis_config=None):
        """
        Initialize cache.
        
        Args:
            redis_config: Redis configuration (optional)
        """
        self.redis_config = redis_config
        self._memory_cache: Dict[str, Any] = {}
        self._enabled = False
        
        # Future: Initialize Redis connection if config provided
        # if redis_config and redis_config.enabled:
        #     import redis.asyncio as redis
        #     self.redis_client = redis.Redis(...)
        #     self._enabled = True
    
    async def get(self, key: str) -> Optional[Any]:
        """Get value from cache."""
        if not self._enabled:
            return self._memory_cache.get(key)
        # Future: Redis implementation
        return None
    
    async def set(self, key: str, value: Any, ttl: Optional[int] = None):
        """Set value in cache."""
        if not self._enabled:
            self._memory_cache[key] = value
        # Future: Redis implementation with TTL
    
    async def delete(self, key: str):
        """Delete key from cache."""
        if not self._enabled:
            self._memory_cache.pop(key, None)
        # Future: Redis implementation
    
    async def clear(self):
        """Clear all cache."""
        if not self._enabled:
            self._memory_cache.clear()
        # Future: Redis implementation

